Split keyword evaluation into 3 phases
Lookps need to be done first, as they may result in more keywords
being inserted into the object. After all lookups are done, it's
possible that we only need one more phase, but it'd covenient to
have an extensible framework, we can always reduce the number of
phases, but for now 3 is a good number.