Reverse bad decision on handling of stale state
As `(*BranchLocator).value` is a pointer to a subtree, it tracks
stale subtrees. We can find other ways to fix this, perhaps rewriting
the package in a way that treats the tree as an immutable entity may
be a little hard to do in Go (due to the lack of read-only constructs),
but we maybe able to call modifiers on `(*BranchLocator).value`  instead
of `(*Converter).tree`, and reducing reducing sub-tree reference count
would probably help too.